rm types
